Pulled Crypto API calls into seperate module.
